chore: Extract validation functions to util folder
